ailabspro/files/extracted/v28-claude-chrome-video/claude-chrome-video/requirements.txt
selenium>=4.15.0
Pillow>=10.0.0
webdriver-manager>=4.0.0